+++ src/main/configyaml.cpp
@@ -1,10 +1,8 @@
-/**
- * Reads configuration file, and allows application to be configured. 
- */
+/*****************************************************************
+ * Reads configuration from YAML file.
+ *****************************************************************/
 
-
-#include "ConfigReader.hpp"
-
+#include "configyaml.hpp"
 
 static inline YAML::Node retrieve_node(string key, YAML::Node node) 
 {
@@ -12,11 +10,10 @@ static inline YAML::Node retrieve_node(string key, YAML::Node node)
     return rv;
 }
 
-
 /**
  * Load configuration from input YAML file. 
  */
-void ConfigReader::load_file(string config_file) {
+void ConfigYamlReader::load_file(string config_file) {
     Logger::info("reading configuration file: " + config_file);
     YAML::Node config = YAML::LoadFile(trim(config_file));
 
@@ -36,37 +33,4 @@ void ConfigReader::load_file(string config_file) {
         _hidden_nodes  = retrieve_node("hidden", node)["nodes"].as<float>();
         _hidden_layers = retrieve_node("hidden", node)["layers"].as<float>();
     }
-}
-
-
- unsigned int ConfigReader::hidden_layers(const string nid) {
-    return _hidden_layers;
- }
-
-unsigned int ConfigReader::hidden_nodes(const string nid) {
-    return _hidden_nodes;
-}
-
-unsigned int  ConfigReader::input_nodes(const string nid) {
-    return _input_nodes;
-}
-
-unsigned int ConfigReader::output_nodes(const string nid) {
-    return _output_nodes;
-}
-
-float ConfigReader::nn_inital_weight(const string nid) {
-    return _nn_inital_weight;
-}
-
-float ConfigReader::nn_success(const string nid) {
-    return _nn_success;
-}
-
-float ConfigReader::nn_learning_rate(const string nid) {
-    return _nn_learning_rate;
-}
-
-float ConfigReader::nn_momentum(const string nid) {
-    return _nn_momentum;
 }
