+++ src/main/nn.hpp
@@ -15,7 +15,7 @@
 #include "Layer.hpp"
 #include "Logger.hpp"
 #include "TrainingSet.hpp"
-#include "ConfigReader.hpp"
+#include "config.hpp"
 
 using namespace std;
